Fixes a bug where the connection would timeout.
In NodeJS 0.10.3 the socket connect event would never fire. Which meant the smtp server greeting would not be sent to the client and the smtp session would not progress. As soon as the server creates the socket you can read from/write to it, you dont need to wait for a connect event, which in 0.10.3 doesnt fire anyway.